nucleotide-count: update to v1.2.0
https://github.com/exercism/problem-specifications/blob/bbdcaee23c6ca53018a7d360279f705de66df228/exercises/nucleotide-count/canonical-data.json
