+++ exercises/nucleotide-count/.meta/src/reference/kotlin/Dna.kt
@@ -1,9 +1,10 @@
-class DNA(strand: String) {
+class Dna(strand: String) {
+
     init {
         require(strand.matches(Regex("^[$nucleotides]*$")), { "DNA sequence contains invalid nucleotides sequence." })
     }
 
-    companion object DNA {
+    companion object Dna {
         val nucleotides = "ATCG"
         val emptyNucleotideCounts = nucleotides.map { it to 0 }.toMap()
     }
@@ -12,10 +13,4 @@ class DNA(strand: String) {
         emptyNucleotideCounts + strand.groupBy { it }.mapValues { it.value.size }
     }
 
-    fun count(nucleotide: Char): Int {
-        require(nucleotide in nucleotides, { "$nucleotide is not a nucleotide" })
-
-        return nucleotideCounts[nucleotide] ?: 0
-    }
 }
-
